% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dynConfiRIDP-package.R
\docType{package}
\name{dynConfiRIDP-package}
\alias{dynConfiRIDP}
\alias{dynConfiRIDP-package}
\title{dynConfiRIDP: Dynamic Models for Confidence and Response Time Distributions}
\description{
Provides response time and confidence distributions (density/PDF) for
the following models: dynaViTE, dynWEV, 2DSD, IRM and PCRM.
The package also provides a formula-based interface for fitting
these models to data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/SeHellmann/dynConfiR_idp}
  \item \url{https://sehellmann.github.io/dynConfiR_idp/}
  \item Report bugs at \url{https://github.com/SeHellmann/dynConfiR_idp}
}

}
\author{
\strong{Maintainer}: Sebastian Hellmann \email{sebastian.hellmann@ku.de} (\href{https://orcid.org/0000-0002-3621-6343}{ORCID})

Authors:
\itemize{
  \item Manuel Rausch \email{manuel.rausch@ku.de} (\href{https://orcid.org/0000-0002-5805-5544}{ORCID}) [funder]
}

}
\keyword{internal}
